retrace: Move the relay race into its own module.
